refactor: extract transformation registry into dedicated module
Split the 1016-line transformation_service.py into two modules:
- app/services/transformation_service.py — pure DataFrame
  transformation functions (filter, sort, pivot, melt, etc.)
- app/services/transforms_registry.py — TransformationError,
  TransformationSpec, TRANSFORMATION_REGISTRY, resolve_transformation,
  and apply_logged_transformation (log-replay dispatch)

The public API is unchanged: transformation_service.py re-exports
TransformationError, TRANSFORMATION_REGISTRY, apply_logged_transformation,
and resolve_transformation so all existing import paths continue to work.

resolve_transformation uses a lazy function-level import to avoid
circular dependencies between the two modules while preserving the
ability for tests to patch functions on transformation_service.

Zero behaviour change. All 458 backend tests pass.
